refactor(surveys): use shared subpath imports in response pipes
Align survey-response-label and survey-response-severity pipes with the
other two survey pipes touched in this PR (can-take-survey,
survey-action-text), which already import from the @lfx-one/shared
subpath entrypoints instead of the root barrel.
